+++ helloworld-blockchain-netcore/src/main/java/com/xingkaichun/helloworldblockchain/netcore/service/BlockChainForkService.java
@@ -2,7 +2,6 @@ package com.xingkaichun.helloworldblockchain.netcore.service;
 
 import com.xingkaichun.helloworldblockchain.netcore.dto.fork.BlockchainForkBlockDto;
 
-import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -12,12 +11,12 @@ import java.util.List;
  */
 public interface BlockChainForkService {
 
-    boolean isFork(BigInteger blockHeight, String blockHash);
+    boolean isFork(long blockHeight, String blockHash);
 
     /**
      * 获取固定Hash的[小于传入的区块高度]的最大区块高度
      */
-    BigInteger getFixBlockHashMaxBlockHeight(BigInteger blockHeight);
+    long getFixBlockHashMaxBlockHeight(long blockHeight);
 
     void updateBlockchainFork(List<BlockchainForkBlockDto> blockList) ;
     void blockchainForkHandler() ;
